lambda_deployment/requirements-local.txt
# Local Development Dependencies
flask==2.3.3
flask-cors==4.0.0
boto3==1.34.0
python-dotenv==1.0.0

# AWS SDK Dependencies  
botocore==1.34.0

# Development Tools
pytest==7.4.3
requests==2.31.0

# Data Processing
pandas==2.1.4

# Type Hints
typing-extensions==4.8.0